// confug/build/types/types.ts
export interface BuildPaths {
  entry: string
  html: string
  output: string
  src:string
  public: string
}

export type BuildMode = 'production' | 'development'
export type BuildPlatform = 'modile' | 'desktop'

export interface BuildOptions {
  port: number
  paths: BuildPaths
  mode: BuildMode
  analyzer?: boolean
  platform: BuildPlatform
}